+++ src/packages/relations/relation-types/collection/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_RELATION_TYPE_COLLECTION_ALIAS = 'Umb.Collection.RelationType';
