+++ src/packages/relations/relation-types/collection/index.ts
@@ -1,2 +1,2 @@
 export { UmbRelationTypeCollectionRepository } from './repository/index.js';
-export { UMB_RELATION_TYPE_COLLECTION_ALIAS } from './manifests.js';
+export { UMB_RELATION_TYPE_COLLECTION_ALIAS } from './constants.js';
