+++ src/packages/relations/relation-types/collection/manifests.ts
@@ -1,9 +1,8 @@
+import { UMB_RELATION_TYPE_COLLECTION_ALIAS } from './constants.js';
 import { UMB_RELATION_TYPE_COLLECTION_REPOSITORY_ALIAS } from './repository/index.js';
 import { manifests as collectionRepositoryManifests } from './repository/manifests.js';
 import { manifests as collectionViewManifests } from './views/manifests.js';
 import type { ManifestCollection, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_RELATION_TYPE_COLLECTION_ALIAS = 'Umb.Collection.RelationType';
 
 const collectionManifest: ManifestCollection = {
 	type: 'collection',
