+++ src/packages/relations/relation-types/collection/repository/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_RELATION_TYPE_COLLECTION_REPOSITORY_ALIAS = 'Umb.Repository.RelationType.Collection';
