+++ src/packages/relations/relation-types/collection/repository/index.ts
@@ -1,2 +1,2 @@
-export { UMB_RELATION_TYPE_COLLECTION_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_RELATION_TYPE_COLLECTION_REPOSITORY_ALIAS } from './constants.js';
 export { UmbRelationTypeCollectionRepository } from './relation-type-collection.repository.js';
