+++ src/packages/relations/relation-types/repository/detail/constants.ts
@@ -1,0 +1,2 @@
+export const UMB_RELATION_TYPE_DETAIL_REPOSITORY_ALIAS = 'Umb.Repository.RelationType.Detail';
+export const UMB_RELATION_TYPE_DETAIL_STORE_ALIAS = 'Umb.Store.RelationType.Detail';
