+++ src/packages/relations/relation-types/repository/detail/index.ts
@@ -1,2 +1,2 @@
 export { UmbRelationTypeDetailRepository } from './relation-type-detail.repository.js';
-export { UMB_RELATION_TYPE_DETAIL_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_RELATION_TYPE_DETAIL_REPOSITORY_ALIAS } from './constants.js';
